updated the file structure
