+++ Bipin/day1/002_anagram/go.mod
@@ -1,0 +1,3 @@
+module anagram
+
+go 1.22.5
