+++ Bipin/day1/002_anagram/main.go
@@ -1,0 +1,9 @@
+package main
+
+import "fmt"
+
+func main() {
+	// Example usage: anagram
+	fmt.Println(isAnagram("anagram", "nagaram")) // true
+	fmt.Println(isAnagram("rat", "car"))         // false
+}
