+++ Bipin/day1/003_two_sum/go.mod
@@ -1,0 +1,3 @@
+module two_sum
+
+go 1.22.5
