+++ Bipin/day1/003_two_sum/main.go
@@ -1,0 +1,9 @@
+package main
+
+import "fmt"
+
+func main() {
+	// Example usage: two_sum
+	fmt.Println(twoSum([]int{2, 7, 11, 15}, 9)) // [0, 1]
+	fmt.Println(twoSum([]int{3, 2, 4}, 6))      // [1, 2]
+}
