+++ Bipin/day1/003_two_sum/problem.md
@@ -1,0 +1,12 @@
+# Two Sum
+Given an array of integers nums and an integer target, return indices of the two numbers such that they add up to target.
+
+# Example:
+Input: nums = [2,7,11,15], target = 9
+Output: [0,1]
+
+# Unit Tests:
+1. Test with a valid input where a solution exists
+2. Test with no valid solution
+3. Test with multiple valid solutions (return any one)
+4. Test with an array of length 2
