--- Bipin/day1/go.mod
@@ -1,3 +1,0 @@
-module reverse-string
-
-go 1.22.5
